+++ tests/NodaMoney.Tests/Serialization/TestDataGenerator.cs
@@ -14,13 +14,13 @@ public class ValidJsonV1TestData : TheoryData<string, Money>
         var money = new Money(234.25m, Currency.FromCode("EUR"));
 
         Add("""{ "Amount": 234.25, "Currency": "EUR" }""", money); // PascalCase, Amount as number
-        Add("""{ "Currency": "EUR", "Amount": 234.25 }""", money); // PascalCase, Amount as number, Reversed memb"ers
-        Add("""{ "Amount": "234.25", "Currency": "EUR" }""", money); // PascalCase, Amount as str"ing
-        Add("""{ "Currency": "EUR", "Amount": "234.25" }""", money); // PascalCase, Amount as string, Reversed memb"ers
+        Add("""{ "Currency": "EUR", "Amount": 234.25 }""", money); // PascalCase, Amount as number, Reversed members
+        Add("""{ "Amount": "234.25", "Currency": "EUR" }""", money); // PascalCase, Amount as string
+        Add("""{ "Currency": "EUR", "Amount": "234.25" }""", money); // PascalCase, Amount as string, Reversed members
         Add("""{ "amount": 234.25, "currency": "EUR" }""", money); // camelCase, Amount as num"ber
-        Add("""{ "currency": "EUR", "amount": 234.25 }""", money); // camelCase, Amount as number, Reversed memb"ers
-        Add("""{ "amount": "234.25", "currency": "EUR" }""", money); // camelCase, Amount as str"ing
-        Add("""{ "currency": "EUR", "amount": "234.25" }""", money); // camelCase, Amount as string, Reversed memb"ers
+        Add("""{ "currency": "EUR", "amount": 234.25 }""", money); // camelCase, Amount as number, Reversed members
+        Add("""{ "amount": "234.25", "currency": "EUR" }""", money); // camelCase, Amount as string
+        Add("""{ "currency": "EUR", "amount": "234.25" }""", money); // camelCase, Amount as string, Reversed members
 
         // // Members no quotation marks (NOT valid json)
         // Add("{ Amount: 234.25, Currency: \"EUR\" }", money); // PascalCase, Amount as number